Move AspNet Span Helper method to shared class
